+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- New `transform` parameter that accepts a dictionnary of field-function items where each
+  function transforms the input value that will be displayed. Fixes: Issue #10
+
 ## [0.0.4] - 2021/04/01
 ### Changed
 - The demo notebook can now be run on Google Colab